pyproject.toml
[project]
name = "unzipper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.4",
    "pillow>=11.3.0",
    "py7zr>=1.0.0",
    "python-dotenv>=1.1.1",
    "rarfile>=4.2",
]
